fix typo

+++ build-image.sh
@@ -83,6 +83,7 @@ fi
 # chroot into target
 mount --bind ${BUILD_PATH} ${BUILD_PATH}
 # arch-chroot ${BUILD_PATH} /bin/bash -c "cd / && /all-install.sh"
+
 # 重试次数
 MAX_RETRIES=3
 RETRY_COUNT=0
@@ -98,6 +99,10 @@ while [ ${RETRY_COUNT} -lt ${MAX_RETRIES} ]; do
 	break
 done
 set -e
+if [ ${RETRY_COUNT} -eq ${MAX_RETRIES} ]; then
+	echo ">>>>>> All install failed after ${MAX_RETRIES} attempts. Stopping..."
+	exit -1
+fi
 
 rm ${BUILD_PATH}/all-install.sh
 rm ${BUILD_PATH}/postinstall
